requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
alembic
aiogram==3.4